Web BI/requirements.txt
# Core dependencies
streamlit==1.40.2
duckdb==1.1.3
pandas==2.2.3
numpy==1.26.4
plotly==5.24.1

# Data processing
openpyxl==3.1.5  # Excel export
python-dateutil==2.9.0

# Visualization
# Pin Altair <5 to satisfy streamlit-aggrid 0.3.5 constraint
altair==4.2.2
matplotlib==3.9.2
seaborn==0.13.2

# Utilities
pyyaml==6.0.2
python-dotenv==1.0.1

# Caching and performance
streamlit-aggrid==0.3.5  # Advanced grid components